fix(208): cast DEPLOY_DAY_BOUNDARY to uint32 in GameTimeLib
ContractAddresses.DEPLOY_DAY_BOUNDARY is uint48 (user-managed file).
Local cast avoids type promotion error without modifying ContractAddresses.sol.

+++ contracts/libraries/GameTimeLib.sol
@@ -30,6 +30,6 @@ library GameTimeLib {
      */
     function currentDayIndexAt(uint48 ts) internal pure returns (uint32) {
         uint32 currentDayBoundary = uint32((ts - JACKPOT_RESET_TIME) / 1 days);
-        return currentDayBoundary - ContractAddresses.DEPLOY_DAY_BOUNDARY + 1;
+        return currentDayBoundary - uint32(ContractAddresses.DEPLOY_DAY_BOUNDARY) + 1;
     }
 }
